refactor(core): move FitState/TuningState to facade-adjacent _model_state (H-0084, #171)
