+++ tests/test_core/test_mixin_state_isolation.py
@@ -29,8 +29,8 @@ import pandas as pd
 import pytest
 
 from lizyml import Model
+from lizyml.core._model_state import TuningState
 from lizyml.core.exceptions import ErrorCode, LizyMLError
-from lizyml.core.types.fit_state import TuningState
 from tests._helpers import make_config, make_regression_df
 
 # ---------------------------------------------------------------------------
